refactor: extract schema utilities to agno.utils.schema
Move type annotation utilities from function.py to a shared module:
- unwrap_annotation, is_union, annotation_reaches, annotation_binds
- is_framework_typed, is_bare_media_typed, is_schema_excluded
- FRAMEWORK_INJECTED_PARAMS, AGNO_INJECTED_PARAMS, identity_injected_types

This allows MCP to reuse the same logic for hiding framework params.
